LayoutTests/media/video-poster-delayed-expected.txt


Video loaded

EXPECTED (video.clientWidth == '320') OK
EXPECTED (video.clientHeight == '240') OK
EXPECTED (video.videoWidth == '320') OK
EXPECTED (video.videoHeight == '240') OK

Poster loaded

EXPECTED (video.clientWidth == '320') OK
EXPECTED (video.clientHeight == '240') OK
EXPECTED (video.videoWidth == '320') OK
EXPECTED (video.videoHeight == '240') OK
END OF TEST
